EgoMotion/requirements.txt
torch
numpy
pyyaml
torchvision
opencv-python
matplotlib
